refactor: WaitingService 이름 변경 및 역할 재분배
