+++ src/main/java/roomescape/controller/WaitingController.java
@@ -12,38 +12,38 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.LoginMemberId;
 import roomescape.service.reservation.dto.ReservationResponse;
-import roomescape.service.waiting.WaitingCommonService;
-import roomescape.service.waiting.WaitingCreateService;
+import roomescape.service.waiting.WaitingCommandService;
+import roomescape.service.waiting.WaitingQueryService;
 import roomescape.service.waiting.dto.WaitingRequest;
 
 @RestController
 @RequestMapping("/waitings")
 public class WaitingController {
 
-    private final WaitingCommonService waitingCommonService;
-    private final WaitingCreateService waitingCreateService;
+    private final WaitingQueryService waitingQueryService;
+    private final WaitingCommandService waitingCommandService;
 
-    public WaitingController(WaitingCommonService waitingCommonService, WaitingCreateService waitingCreateService) {
-        this.waitingCommonService = waitingCommonService;
-        this.waitingCreateService = waitingCreateService;
+    public WaitingController(WaitingQueryService waitingQueryService, WaitingCommandService waitingCommandService) {
+        this.waitingQueryService = waitingQueryService;
+        this.waitingCommandService = waitingCommandService;
     }
 
     @PostMapping
     public ResponseEntity<ReservationResponse> createWaiting(
         @RequestBody WaitingRequest waitingRequest,
         @LoginMemberId Long memberId) {
-        ReservationResponse response = waitingCreateService.createWaiting(waitingRequest, memberId);
+        ReservationResponse response = waitingCommandService.createWaiting(waitingRequest, memberId);
         return ResponseEntity.created(URI.create("/waitings/" + response.id())).body(response);
     }
 
     @GetMapping
     public List<ReservationResponse> findAll() {
-        return waitingCommonService.findAll();
+        return waitingQueryService.findAll();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteWaiting(@PathVariable("id") long reservationId, @LoginMemberId long memberId) {
-        waitingCommonService.deleteWaitingById(reservationId, memberId);
+        waitingCommandService.deleteWaitingById(reservationId, memberId);
         return ResponseEntity.noContent().build();
     }
 }
