+++ src/test/java/roomescape/service/waiting/WaitingCommandServiceTest.java
@@ -26,16 +26,17 @@ import roomescape.domain.schedule.ReservationTimeRepository;
 import roomescape.domain.schedule.Schedule;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.ThemeRepository;
+import roomescape.exception.ForbiddenException;
 import roomescape.exception.InvalidReservationException;
 import roomescape.service.reservation.dto.ReservationResponse;
 import roomescape.service.waiting.dto.WaitingRequest;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @Sql("/truncate.sql")
-class WaitingCreateServiceTest {
+class WaitingCommandServiceTest {
 
     @Autowired
-    private WaitingCreateService waitingCreateService;
+    private WaitingCommandService waitingCommandService;
     @Autowired
     private ReservationRepository reservationRepository;
     @Autowired
@@ -71,7 +72,7 @@ class WaitingCreateServiceTest {
         reservationRepository.save(new Reservation(anotherMember, reservationDetail, ReservationStatus.RESERVED));
 
         //when
-        ReservationResponse result = waitingCreateService.createWaiting(waitingRequest, member.getId());
+        ReservationResponse result = waitingCommandService.createWaiting(waitingRequest, member.getId());
 
         //then
         assertAll(
@@ -91,9 +92,34 @@ class WaitingCreateServiceTest {
         reservationRepository.save(new Reservation(member, reservationDetail, ReservationStatus.RESERVED));
 
         //when & then
-        assertThatThrownBy(() -> waitingCreateService.createWaiting(waitingRequest, member.getId()))
+        assertThatThrownBy(() -> waitingCommandService.createWaiting(waitingRequest, member.getId()))
             .isInstanceOf(InvalidReservationException.class)
             .hasMessage("이미 예약(대기) 상태입니다.");
     }
 
+    @DisplayName("사용자가 예약 대기를 삭제하려고 할 때 예약으로 바뀌었다면 예외가 발생한다.")
+    @Test
+    void cannotDeleteWaitingByIdIfReserved() {
+        //given
+        Reservation reservation = new Reservation(member, reservationDetail, ReservationStatus.RESERVED);
+        Reservation target = reservationRepository.save(reservation);
+
+        //when
+        assertThatThrownBy(() -> waitingCommandService.deleteWaitingById(target.getId(), member.getId()))
+            .isInstanceOf(InvalidReservationException.class)
+            .hasMessage("예약은 삭제할 수 없습니다. 관리자에게 문의해주세요.");
+    }
+
+    @DisplayName("사용자가 본인 외 예약 대기를 삭제하려고 하면 예외가 발생한다.")
+    @Test
+    void cannotDeleteWaitingByIdIfNotOwner() {
+        //given
+        Reservation reservation = new Reservation(member, reservationDetail, ReservationStatus.WAITING);
+        Reservation target = reservationRepository.save(reservation);
+
+        //when
+        assertThatThrownBy(() -> waitingCommandService.deleteWaitingById(target.getId(), anotherMember.getId()))
+            .isInstanceOf(ForbiddenException.class)
+            .hasMessage("예약 대기를 삭제할 권한이 없습니다.");
+    }
 }
